+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/_utils.py
@@ -375,6 +375,54 @@ def get_validate_platform(cmd, os_type, platform):
     return platform_os, platform_arch, platform_variant
 
 
+def get_custom_registry_credentials(cmd, auth_mode=None, login_server=None, username=None, password=None):
+    """Get the credential object from the input
+    :param str auth_mode: The login mode for the source registry
+    :param str login_server: The login server of custom registry
+    :param str username: The username for custom registry
+    :param str password: The password for custom registry
+    """
+    source_registry_credentials = None
+    if auth_mode:
+        SourceRegistryCredentials = cmd.get_models('SourceRegistryCredentials')
+        source_registry_credentials = SourceRegistryCredentials(login_mode=auth_mode)
+
+    custom_registries = None
+    if login_server:
+        CustomRegistryCredentials, SecretObject, SecretObjectType = cmd.get_models(
+            'CustomRegistryCredentials',
+            'SecretObject',
+            'SecretObjectType'
+        )
+
+        # if Null username and password, then have it removed
+        custom_reg_credential = None
+        if (username and not password) or (password and not username):
+            raise CLIError("Please provide both username and password.")
+        elif username and password:
+            custom_reg_credential = CustomRegistryCredentials(
+                user_name=SecretObject(
+                    type=SecretObjectType.opaque,
+                    value=username
+                ),
+                password=SecretObject(
+                    type=SecretObjectType.opaque,
+                    value=password
+                )
+            )
+        elif username is not None or password is not None:
+            # case where user passes empty strings: -u '' -p ''
+            raise CLIError("username and password cannot be empty")
+
+        custom_registries = {login_server: custom_reg_credential}
+
+    Credentials = cmd.get_models('Credentials')
+    return Credentials(
+        source_registry=source_registry_credentials,
+        custom_registries=custom_registries
+    )
+
+
 class ResourceNotFound(CLIError):
     """For exceptions that a resource couldn't be found in user's subscription
     """