+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/_help.py
@@ -55,6 +55,11 @@ helps['ad sp create-for-rbac'] = """
           text: az ad sp create-for-rbac --keyvault MyVault --cert CertName
     """
 
+helps['ad sp update'] = """
+    type: command
+    short-summary: update a service principal
+"""
+
 helps['ad sp credential'] = """
     type: group
     short-summary: manage a service principal's credentials.
@@ -78,7 +83,7 @@ helps['ad sp credential reset'] = """
     long-summary: Use upon expiration of the service principal's credentials, or in the event that login credentials are lost.
     parameters:
         - name: --name -n
-          short-summary: Name or app URI for the credential.
+          short-summary: Name or app ID of the service principal.
         - name: --password -p
           short-summary: The password used to log in.
           long-summary: If not present and `--cert` is not specified, a random password will be generated.
@@ -195,10 +200,16 @@ helps['ad app permission'] = """
 """
 helps['ad app permission grant'] = """
     type: command
-    short-summary: Grant the app an API permission
+    short-summary: Grant the app an API Delegated permissions
+    long-summary: for Application permissions, please use "ad app permission admin-consent"
     examples:
         - name: Grant a native application with permissions to access an existing API with TTL of 2 years
           text: az ad app permission grant --id e042ec79-34cd-498f-9d9f-1234234 --api a0322f79-57df-498f-9d9f-12678 --expires 2
+"""
+helps['ad app permission admin-consent'] = """
+    type: command
+    short-summary: grant Application & Delegated permissions through admin-consent.
+    long-summary: you must login as a directory administrator
 """
 helps['ad app permission list'] = """
     type: command
@@ -222,6 +233,14 @@ helps['ad app permission delete'] = """
         - name: remove an AAD graph permission
           text: az ad app permission delete --id eeba0b46-78e5-4a1a-a1aa-cafe6c123456 --api 00000002-0000-0000-c000-000000000000
 """
+helps['ad app permission list-grants'] = """
+    type: command
+    short-summary: List Oauth2 permission grants
+    examples:
+        - name: list oauth2 permissions granted to the service principal
+          text: az ad app permission list-grants --id e042ec79-34cd-498f-9d9f-1234234123456
+"""
+
 helps['ad app credential'] = """
     type: group
     short-summary: manage an application's password or certificate credentials