+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/_params.py
@@ -6,6 +6,7 @@
 # pylint: disable=line-too-long
 
 from knack.arguments import CLIArgumentType
+from azure.graphrbac.models import ConsentType
 
 from azure.cli.core.commands.parameters import get_enum_type, get_three_state_flag, get_location_type, tags_type
 from azure.cli.core.commands.validators import validate_file_or_dict
@@ -55,8 +56,13 @@ def load_arguments(self, _):
     with self.argument_context('ad app permission') as c:
         c.argument('api_permissions', nargs='+', help='space seperated list of <resource-access-id>=<type>')
         c.argument('expires', help='Expiry date for the permissions in years. e.g. 1, 2 or "never"')
-        c.argument('scope', help='oauth scope')
+        c.argument('scope', help='Specifies the value of the scope claim that the resource application should expect in the OAuth 2.0 access token, e.g. User.Read')
         c.argument('api', help='the target API to access')
+        c.argument('consent_type', arg_type=get_enum_type(ConsentType), default=ConsentType.all_principals.value,
+                   help="Indicates if consent was provided by the administrator (on behalf of the organization) or by an individual.")
+        c.argument('principal_id', help='If --consent-type is "Principal", this argument specifies the object of the user that granted consent and applies only for that user.')
+        c.argument('show_resource_name', options_list=['--show-resource-name', '-r'],
+                   arg_type=get_three_state_flag(), help="show resource's display name")
 
     with self.argument_context('ad app permission list') as c:
         c.argument('identifier', options_list=['--id'], help='identifier uri, application id, or object id of the associated application')
@@ -81,7 +87,9 @@ def load_arguments(self, _):
         with self.argument_context('ad sp {}'.format(item)) as c:
             c.argument('name', name_arg_type)
             c.argument('cert', arg_group='Credential', validator=validate_cert)
-            c.argument('years', type=int, default=None, arg_group='Credential')
+            c.argument('years', default=None, arg_group='Credential')
+            c.argument('end_date', default=None, arg_group='Credential',
+                       help="Finer grain of expiry time if '--years' is insufficient, e.g. '2020-12-31T11:59:59+00:00' or '2299-12-31'")
             c.argument('create_cert', action='store_true', arg_group='Credential')
             c.argument('keyvault', arg_group='Credential')
             c.argument('append', action='store_true', help='Append the new credential instead of overwriting.')
@@ -89,7 +97,7 @@ def load_arguments(self, _):
 
     with self.argument_context('ad sp create-for-rbac') as c:
         c.argument('password', options_list=['--password', '-p'], arg_group='Credential',
-                   deprecate_info=c.deprecate(expiration='2.1.0', hide=False), help="If missing, CLI will generate a strong password")
+                   deprecate_info=c.deprecate(hide=False), help="If missing, CLI will generate a strong password")
 
     with self.argument_context('ad sp credential reset') as c:
         c.argument('password', options_list=['--password', '-p'], arg_group='Credential',