+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -6,6 +6,8 @@ Release History
 2.2.2
 +++++
 * Fix redundant sources in image import.
+* Add '--auth-mode' for 'az acr build', 'az acr run', 'az acr task create', and 'az acr task update' commands.
+* Add 'acr task credential' command group for managing credentials for a Task.
 
 2.2.1
 +++++