+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -8,6 +8,7 @@ Release History
 * Fix redundant sources in image import.
 * Add '--auth-mode' for 'az acr build', 'az acr run', 'az acr task create', and 'az acr task update' commands.
 * Add 'acr task credential' command group for managing credentials for a Task.
+* Add '--no-wait' for 'az acr build' command.
 
 2.2.1
 +++++